Q&A CHATBOT/requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit
langchain
langchain-community
langchain-core
python-dotenv
ollama
